Moved collision and constraints out of core
