+++ Collision/Obstacle.hpp
